backend/requirements.txt
# Backend framework
fastapi
uvicorn[standard]

# Async HTTP client
httpx

# Data validation and settings
pydantic>=2
pydantic-settings

# HTML parsing and boilerplate removal
beautifulsoup4
readability-lxml
trafilatura

# Convert HTML -> Markdown (helps with chunking)
markdownify

# LLM integration
langgraph
langchain
langchain-community
langchain-openai
langchain-google-genai
langchain-anthropic

# Anthropic API (optional if using Claude agents)
anthropic

# Optional: Token management for chunk size control
tiktoken

# Optional: Environment variable management
python-dotenv
